+++ src/pages/Shopping/Shopping.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
+import Shoppinglist from './Shoppinglist';
 import './Shopping.scss';
 
 function Shopping() {
@@ -11,9 +12,6 @@ function Shopping() {
       .then(response => response.json())
       .then(data => setItemInfo(data));
   }, []);
-
-  //상품리스트 삭제
-  const removeItem = id => setItemInfo(itemInfo.filter(item => item.id !== id));
 
   //상품 총 가격 계산
   let totalprice = () => {
@@ -62,19 +60,6 @@ function Shopping() {
       qty: 1,
     },
   ];
-  //수량 선택창
-  /*const qtySelect = [...Array(8).keys()].map(key => key + 1);*/
-
-  const [qtyOpen, setQtyOpen] = useState(false);
-
-  //수량 선택창 또 다른 방법
-  /* const QtyInput = () => {
-    const arr = [];
-    for (let i = 1; i < 9; i++) {
-      arr.push(<button>{i}</button>);
-    }
-    return arr; 
-  };*/
 
   return (
     <div className="shopping">
@@ -90,32 +75,12 @@ function Shopping() {
         </div>
         {itemInfo.map((itemdata, i) => {
           return (
-            <div className="buyitems" key={itemdata.id}>
-              <div>
-                <p>{itemdata.koName}</p>
-                <p>{itemdata.enName}</p>
-                <p>{itemdata.size}</p>
-              </div>
-              <div>₩{itemdata.price}</div>
-              <div className="qtyarea">
-                <button
-                  onClick={() => {
-                    setQtyOpen(true);
-                  }}
-                >
-                  {itemdata.qty} ⋁
-                </button>
-              </div>
-              <div className="totalprice">
-                <p>₩{itemdata.pricenum * itemdata.qty}</p>
-                <p
-                  className="deleteitem"
-                  onClick={() => removeItem(itemdata.id)}
-                >
-                  삭제
-                </p>
-              </div>
-            </div>
+            <Shoppinglist
+              itemdata={itemdata}
+              itemInfo={itemInfo}
+              setItemInfo={setItemInfo}
+              key={itemdata.id}
+            />
           );
         })}
         <div className="freeitem">
@@ -126,8 +91,8 @@ function Shopping() {
         </div>
         <p className="totalpay">합계 ₩{totalprice()}</p>
         <div className="buttonarea">
-          <Link to="/">
-            <div className="continueshop">쇼핑 계속하기</div>
+          <Link to="/" className="continueshop">
+            쇼핑 계속하기
           </Link>
           <Link to="/">
             <button className="buybutton">결제하기</button>
